+++ src/openepd/model/specs/range/cmu.py
@@ -21,7 +21,7 @@ __all__ = ("CMURangeV1",)
 from openepd.compat.pydantic import pyd
 from openepd.model.common import RangeFloat
 from openepd.model.specs.base import BaseOpenEpdHierarchicalSpec
-from openepd.model.specs.generated.enums import CmuBlockType, CmuWeightClassification
+from openepd.model.specs.enums import CmuBlockType, CmuWeightClassification
 from openepd.model.validation.quantity import AmountRangeGWP, AmountRangePressureMpa