+++ src/openepd/model/specs/range/mechanical_insulation.py
@@ -21,7 +21,7 @@ __all__ = ("MechanicalInsulationRangeV1",)
 from openepd.compat.pydantic import pyd
 from openepd.model.common import RangeFloat
 from openepd.model.specs.base import BaseOpenEpdHierarchicalSpec
-from openepd.model.specs.generated.enums import InsulatingMaterial, InsulationIntendedApplication
+from openepd.model.specs.enums import InsulatingMaterial, InsulationIntendedApplication
 from openepd.model.validation.quantity import AmountRangeLengthMm