+++ src/openepd/model/specs/range/thermal_moisture_protection.py
@@ -44,7 +44,7 @@ __all__ = (
 from openepd.compat.pydantic import pyd
 from openepd.model.common import RangeFloat, RangeRatioFloat
 from openepd.model.specs.base import BaseOpenEpdHierarchicalSpec
-from openepd.model.specs.generated.enums import (
+from openepd.model.specs.enums import (
     FoamType,
     InsulatingMaterial,
     InsulationIntendedApplication,