+++ src/openepd/model/specs/singular/aggregates.py
@@ -16,7 +16,7 @@
 from openepd.compat.pydantic import pyd
 from openepd.model.base import BaseOpenEpdSchema
 from openepd.model.specs.base import BaseOpenEpdHierarchicalSpec
-from openepd.model.specs.generated.enums import AggregateGradation, AggregateWeightClassification
+from openepd.model.specs.enums import AggregateGradation, AggregateWeightClassification
 from openepd.model.validation.numbers import RatioFloat
 from openepd.model.validation.quantity import LengthMmStr