+++ src/openepd/model/specs/singular/electrical.py
@@ -15,8 +15,8 @@
 #
 from openepd.compat.pydantic import pyd
 from openepd.model.specs.base import BaseOpenEpdHierarchicalSpec
-from openepd.model.specs.generated.enums import CableTraysMaterial, EnergySource, RacewaysMaterial
-from openepd.model.specs.generated.mixins.conduit_mixin import ConduitMixin
+from openepd.model.specs.enums import CableTraysMaterial, EnergySource, RacewaysMaterial
+from openepd.model.specs.singular.mixins.conduit_mixin import ConduitMixin
 from openepd.model.validation.quantity import (
     ColorTemperatureStr,
     LengthMmStr,