+++ src/openepd/model/specs/singular/manufacturing_inputs.py
@@ -16,7 +16,7 @@
 from openepd.compat.pydantic import pyd
 from openepd.model.specs.base import BaseOpenEpdHierarchicalSpec
 from openepd.model.specs.concrete import Cementitious
-from openepd.model.specs.generated.enums import (
+from openepd.model.specs.enums import (
     AdmixtureEffects,
     CarpetYarnType,
     CementAstmType,