+++ src/openepd/model/specs/singular/mechanical_insulation.py
@@ -15,7 +15,7 @@
 #
 from openepd.compat.pydantic import pyd
 from openepd.model.specs.base import BaseOpenEpdHierarchicalSpec
-from openepd.model.specs.generated.enums import InsulatingMaterial, InsulationIntendedApplication
+from openepd.model.specs.enums import InsulatingMaterial, InsulationIntendedApplication
 from openepd.model.validation.quantity import LengthMmStr