+++ src/openepd/model/tests/test_spec_versions.py
@@ -16,7 +16,7 @@
 import unittest
 
 from openepd.model.specs.base import BaseOpenEpdHierarchicalSpec
-from openepd.model.specs.generated.steel import SteelV1
+from openepd.model.specs.singular.steel import SteelV1
 
 
 class SpecVersionTestCase(unittest.TestCase):